Completely reworked PhysicalCircuit.
May seem like overkill,
but this has a bunch of convenience built-in.
Users do not have to load/create pyGSTi circuits,
making it possible to switch intermediate representations later.
The IsCastable class gives an easy way to define TypeAliases for allowed input,
and gives a cast() function that is copy-friendly,
and enables passing in "raw" inputs that are cast to the correct object.
The docstrings and tutorials are fully up-to-date and passing.