Very incomplete outline of class structure.
Still a lot of thinking to do, but generally I'm imagining two types of objects:
Records, which are a generalized type of state,
and Operations, which take Records and create new ones.
More documentation will come as everything starts to come together,
but wanted to get my initial ideas down first.