Fix all typing issues.
This also includes the removal of the classproperties
stuff I was messing around with.
In the end, I'm trying to just use static typing
instead of the dynamic typing that class properties
were intended to enable.
It has already been worth it though,
as mypy caught several bugs and forced me to clean the interfaces up.