Dask working, but not parallelizing?
I've spent some time optimizing the Dask farming out of shot jobs,
and it seems to be working in that it is faster than serial.
However, it seems to prioritize using only one worker rather than all of them.
Need to understand this behaviour before it is a complete success.