Update damp/dephase tests.
They are now 3Q tests, ensuring that the state of other qubits is not disturbed.
This is in preparation for the Kraus rep for STIM,
where there is some question about embedding/reduced the state space.
I will still leave that for future work though,
since it is not current critical path.